Change -d (delimiter) cli option to take keyword
The lines are parsed using clojure.string/split, which uses a regex.
There are some complications around this:
- the pipe character is a special regex symbol
- depending on if you're defining the regex at compile time or runtime
  the rules for escape characters are different

So either a) the way lines are parsed is changed or b) the delimiter is
a lookup against a known set of regexes. All of the expected options for
field separators are already known, so b) is easy and a) isn't worth the
extra work.